8384080: Shenandoah: Simplify C1 CAS barriers
